Use initsplit to provide machine-specific config via Customize
Rather than creating our own functionality to write an Elisp file
containing settings for a specific machine, we can just use the
Customize system already built into Emacs, along with Initsplit to
save specific settings to a separate custom Elisp file. Right now the
only thing we used this for anyway was setting the default font size,
so we can have the default face settings get saved to this file as
well; in the future we can add more settings.