Remove dark/light theme system
I never really use this anymore - it was an idea I came up with when
using my old laptop outside, because it wasn't bright enough to see in
the sun. These days, I rarely use my laptop outside, and even when I
do, it is bright enough now and has an anti-glare screen as
well. Since I no longer have a need for this, it is pretty much
technical debt, so I'm removing it.