Fix initsplit configuration
What I had there isn't even the correct form - the alist elements take four
sub-elements, with the last two being whether to byte-compile and whether to load the
split custom files at startup, respectively. Also, I had my "two" elements as part of
one element, and had the machine-config.el file as the file for both.

Also, autoloading on initsplit-load will cause us to never load the package - rather,
since we have the hook on init, we should just rely on that to autoload the package.

What was I thinking?