Clean up initsplit use-package declaration
After examining the internals of initsplit.el, it already loads all files marked as
"preload" (the fourth argument in the alist elements) when the feature is
loaded. Therefore, it isn't necessary for us to define our own logic to load
initsplit files, which can clean things up significantly. Additionally, it isn't
necessary for us to load custom-file ourselves; if we set it, initsplit will load it
automatically along with the other files. Note that in order for this to work, both
variables need to be set in the `:init` stage in `use-package`; otherwise, they
will not be set until after initsplit loads and attempts to load files, and our
customizations won't be loaded.

Also, defines find-if as alias for cl-find-if. initsplit.el uses the deprecated
`find-if` (which is no longer defined) from `cl` instead of the newer
`cl-find-if`. This causes some errors to throw when attempting to customize
variables. To fix this temporarily, we define an alias before loading the package,
although ideally the package should be properly patched somehow.