Allow org-roam-dailies-goto-* to work with multiple capture templates
Apparently these functions just work by initiating a capture that doesn't actually
capture anything, so the file can get created if it doesn't exist. This has the
effect of prompting the user for a capture template to use even if the file already
exists, which is really annoying. You can bypass this by setting a `keys` argument
that only works from Elisp, so to fix this I just created some function advice that
calls the original functions with the key for the default template ('d') passed to
this argument.