Create function for splitting a subtask off the current task
Often when I start a task I'll get distracted on something that needs to happen
either before that task or as a subtask of that task that wasn't apparent when I was
originally writing the task TODO item. With this function, I can easily capture a
subtask while clocked into a task, automatically filing it as a sub-entry in that
task, and clock into the new subtask.

Next steps would be to bind it to a key, and also maybe look into some method for
going back to the parent task when done.