Don't use Consult for xref
I actually kind of like the default xref popup that remains persistent; it is useful
when trying to look at references one by one and I need to return to the list of
references. The Consult menu seems intended for jumping to a specific reference that
you can identify and not needing to navigate to another.